# api/openapi.yaml
openapi: 3.1.0
info:
  title: Loom CLI API
  description: |
    REST API for the Loom CLI web UI and daemon. Covers workspace management,
    issue tracking, agent lifecycle, terminal sessions, git operations, fleet
    coordination, monitoring, and observability.
  version: 0.1.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server (post-consolidation single server)
...
security:
  - BearerAuth: []
paths:
...
  /api/config:
    get:
      operationId: getAuthConfig
      tags: [config]
      summary: Auth mode discovery (public, rate-limited)
      security: []
      responses:
        "200":
          description: Auth configuration
          content:
            application/json:
              schema:
                type: object
                properties:
                  mode:
                    type: string
                    enum: [none, api_key, external]
                  auth_url:
                    type: string
                    description: External auth service URL (only when mode=external)
        "429":
          description: Rate limited
          headers:
            Retry-After:
              schema:
                type: integer
...
  /api/client-errors:
    post:
      operationId: reportClientError
      tags: [health]
      summary: Client-side error reporting (rate-limited, 10 req/min/IP)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Error recorded
        "429":
          description: Rate limited

  # ── Stats ───────────────────────────────────────────────────────────────
...
  /api/workspaces/active:
    get:
      operationId: getActiveWorkspace
      tags: [workspaces]
      summary: Get full topology for the default workspace
      responses:
        "200":
          description: Active workspace topology
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceResponse"
        "404":
          description: No default workspace configured
  /api/workspaces:
    get:
      operationId: listWorkspaces
      tags: [workspaces]
      summary: List all workspaces
      responses:
        "200":
          description: Workspace list
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/WorkspaceSummary"
    post:
      operationId: createWorkspace
      tags: [workspaces]
      summary: Create a new workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, path]
              properties:
                name:
                  type: string
                path:
                  type: string
      responses:
        "202":
          description: Workspace creation started (async job)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  job_id:
                    type: string
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/workspaces/{ws}:
    get:
      operationId: getWorkspace
      tags: [workspaces]
      summary: Get workspace topology
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Workspace topology
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceResponse"
        "404":
          description: Workspace not found
    delete:
      operationId: deleteWorkspace
      tags: [workspaces]
      summary: Delete a workspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Workspace deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Workspace not found
  /api/workspaces/{ws}/monitor/status:
    get:
      operationId: getWorkspaceMonitorStatus
      tags: [monitor]
      summary: Full monitor dashboard data for a workspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Dashboard status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitorStatusResponse"
        "404":
          description: Workspace not found
        "503":
          description: Data collection unavailable
  /api/workspaces/jobs/{id}:
    get:
      operationId: getWorkspaceJob
      tags: [workspaces]
      summary: Poll workspace creation job status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [pending, running, completed, failed]
                  error:
                    type: string
        "404":
          description: Job not found
  /api/workspaces/order:
    put:
      operationId: reorderWorkspaces
      tags: [workspaces]
      summary: Set workspace display order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [order]
              properties:
                order:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /api/workspaces/default:
    put:
      operationId: setDefaultWorkspace
      tags: [workspaces]
      summary: Set the default workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workspace]
              properties:
                workspace:
                  type: string
      responses:
        "200":
          description: Default set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
    delete:
      operationId: clearDefaultWorkspace
      tags: [workspaces]
      summary: Clear the default workspace
      responses:
        "200":
          description: Default cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /api/workspaces/{ws}/name:
    patch:
      operationId: renameWorkspace
      tags: [workspaces]
      summary: Rename a workspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceRenameRequest"
      responses:
        "200":
          description: Renamed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/workspaces/{ws}/config/backend:
    get:
      operationId: getWorkspaceBackendConfig
      tags: [config]
      summary: Get workspace-scoped backend configuration
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Backend config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackendConfigResponse"
    patch:
      operationId: patchWorkspaceBackend
      tags: [config]
      summary: Set workspace backend override
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceBackendPatchRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /api/workspaces/{ws}/config/design-format:
    patch:
      operationId: patchWorkspaceDesignFormat
      tags: [config]
      summary: Set the workspace planner design format
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceDesignFormatPatchRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid design format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/editors/open:
    post:
      operationId: openEditor
      tags: [editors]
      summary: Open a file in an external editor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditorOpenRequest"
      responses:
        "200":
          description: Editor launched
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success:
                    type: boolean

  # ── Issue Tabs ──────────────────────────────────────────────────────────
...
  /api/sessions/notify:
    post:
      operationId: notifySessionChange
      tags: [sessions]
      summary: Notify SSE hub of session status change (agent-to-server)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [task_id, session_id, status, workspace_id]
              properties:
                task_id:
                  type: string
                session_id:
                  type: string
                status:
                  type: string
                workspace_id:
                  type: string
      responses:
        "200":
          description: Notification received

  # ── Fleet ───────────────────────────────────────────────────────────────
...
  /api/monitor/status:
    get:
      operationId: getMonitorStatus
      tags: [monitor]
      summary: Full monitor dashboard data
      responses:
        "200":
          description: Dashboard status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitorStatusResponse"
        "503":
          description: Data collection unavailable
  /api/monitor/agents:
    get:
      operationId: getMonitorAgents
      tags: [monitor]
      summary: Agent list with optional workspace grouping
      responses:
        "200":
          description: Agent list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitorAgentsResponse"
        "503":
          description: Data collection unavailable
  /api/monitor/tasks:
    get:
      operationId: getMonitorTasks
      tags: [monitor]
      summary: Task distribution by status
      responses:
        "200":
          description: Task breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitorTasksResponse"
        "503":
          description: Data collection unavailable
  /api/monitor/stats:
    get:
      operationId: getMonitorStats
      tags: [monitor]
      summary: Monitor statistics (distinct from /api/stats issue stats)
      responses:
        "200":
          description: Monitor stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitorStatsResponse"
        "503":
          description: Data collection unavailable
  /api/monitor/sync:
    get:
      operationId: getMonitorSync
      tags: [monitor]
      summary: Git sync status
      responses:
        "200":
          description: Sync status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitorSyncResponse"
        "503":
          description: Data collection unavailable
...
  /api/monitor/usage:
    get:
      operationId: getMonitorUsage
      tags: [monitor]
      summary: Token usage aggregates
      parameters:
        - name: agent
          in: query
          schema:
            type: string
        - name: backend
          in: query
          schema:
            type: string
        - name: epic
          in: query
          schema:
            type: string
        - name: since
          in: query
          description: "YYYY-MM-DD"
          schema:
            type: string
        - name: until
          in: query
          description: "YYYY-MM-DD"
          schema:
            type: string
      responses:
        "200":
          description: Usage data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageResponse"
        "400":
          description: Invalid date parameter
        "503":
          description: Usage store not initialized
...
  /api/observability/events:
    get:
      operationId: getObservabilityEvents
      tags: [observability]
      summary: Paginated event log
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: type
          in: query
          description: Event type filter
          schema:
            type: string
        - name: agent
          in: query
          description: Agent name filter
          schema:
            type: string
        - name: since
          in: query
          description: RFC3339 timestamp filter
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Paginated events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObservabilityEventsResponse"
        "400":
          description: Invalid parameters

  # ── Auth Token ──────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        API key authentication. The key is a 64-character hex string generated
        from 32 random bytes. Configured via `LOOM_AUTH_API_KEY` env var or
        loaded from `~/.loom/api-key`. Supports fallback to `?token=` query
        parameter for WebSocket/EventSource connections.

    WorkerToken:
      type: http
      scheme: bearer
      description: |
        Static token for internal worker API endpoints. Configured via
        `LOOM_WORKER_TOKEN` env var. If not set, worker API returns 503.

    FleetApiKey:
      type: apiKey
      in: header
      name: X-Fleet-API-Key
      description: |
        API key for fleet registration. Configured via `LOOM_FLEET_API_KEY`
        env var. Constant-time comparison with optional per-IP rate limiting.

    FleetJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token for fleet claim/heartbeat endpoints. HMAC-SHA256 signed.
        Configured via `LOOM_FLEET_JWT_KEY` env var. Token expiry: 1 hour.
  parameters:
    WorkspaceId:
      name: ws
      in: path
      required: true
      description: Workspace identifier
      schema:
        type: string
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          const: false
        error:
          type: string
        code:
          type: string
        retryable:
          type: boolean
        details:
          type: object
    MessageResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          const: true
        message:
          type: string
...
    WorkspaceResponse:
      type: object
      required: [id, name, path, repos, groups, agents, workspaces, default_workspace]
      properties:
        id:
          type: string
        name:
          type: string
        path:
          type: string
        repos:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceRepo"
          default: []
        groups:
          type: array
          items:
            type: string
          default: []
        agents:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceAgentInfo"
          default: []
        workspaces:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceSummary"
          default: []
        workspace_order:
          type: array
          items:
            type: string
        default_workspace:
          type: string
        design_format:
          type: string
          enum: [markdown, html]
    WorkspaceSummary:
      type: object
      required: [id, name, path, active, repo_count, is_default]
      properties:
        id:
          type: string
        name:
          type: string
        path:
          type: string
        active:
          type: boolean
        repo_count:
          type: integer
        is_default:
          type: boolean
        backend:
          type: string
...
    WorkspaceRenameRequest:
      type: object
      required: [new_name]
      properties:
        new_name:
          type: string
    WorkspaceBackendPatchRequest:
      type: object
      required: [backend]
      properties:
        backend:
          type: string
    WorkspaceDesignFormatPatchRequest:
      type: object
      required: [design_format]
      properties:
        design_format:
          type: string
          enum: [markdown, html]

    # ── Config Schemas ──────────────────────────────────────────────────
    BackendConfigResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            backend:
              type: string
            source:
              type: string
            available:
              type: array
              items:
                type: string
            agents:
              type: array
              items:
                $ref: "#/components/schemas/AgentBackendOverride"
        error:
          type: string
...
    EditorOpenRequest:
      type: object
      required: [editor_id, path]
      properties:
        editor_id:
          type: string
        path:
          type: string

    # ── Worker Schemas ──────────────────────────────────────────────────
...
    MonitorStatusResponse:
      type: object
      required: [workspace, agents, tasks, needs_planning, ready_to_implement, needs_review, in_progress, in_progress_list, backlog, closed, agent_tasks, stats, sync, timestamp]
      properties:
        workspace:
          $ref: "#/components/schemas/MonitorWorkspaceInfo"
        agents:
          type: array
          items:
            $ref: "#/components/schemas/MonitorAgentStatus"
        tasks:
          $ref: "#/components/schemas/MonitorTaskSummary"
        needs_planning:
          type: array
          items:
            $ref: "#/components/schemas/MonitorTaskInfo"
        ready_to_implement:
          type: array
          items:
            $ref: "#/components/schemas/MonitorTaskInfo"
        needs_review:
          type: array
          items:
            $ref: "#/components/schemas/MonitorTaskInfo"
        in_progress:
          type: array
          items:
            $ref: "#/components/schemas/MonitorTaskInfo"
        in_progress_list:
          type: array
          items:
            $ref: "#/components/schemas/MonitorTaskInfo"
        backlog:
          type: array
          items:
            $ref: "#/components/schemas/MonitorTaskInfo"
        closed:
          type: array
          items:
            $ref: "#/components/schemas/MonitorTaskInfo"
        agent_tasks:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/MonitorTaskInfo"
        stats:
          $ref: "#/components/schemas/MonitorStats"
        sync:
          $ref: "#/components/schemas/MonitorSyncInfo"
        timestamp:
          type: string
          format: date-time
    MonitorAgentsResponse:
      type: object
      required: [workspace, agents, timestamp]
      properties:
        workspace:
          $ref: "#/components/schemas/MonitorWorkspaceInfo"
        agents:
          type: array
          items:
            $ref: "#/components/schemas/MonitorAgentStatus"
        by_workspace:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/MonitorAgentStatus"
        timestamp:
          type: string
          format: date-time
    MonitorTasksResponse:
      type: object
      required: [summary, needs_planning, ready_to_implement, needs_review, in_progress, backlog, closed, timestamp]
      properties:
        summary:
          $ref: "#/components/schemas/MonitorTaskSummary"
        needs_planning:
          type: array
          items:
            $ref: "#/components/schemas/MonitorTaskInfo"
        ready_to_implement:
          type: array
          items:
            $ref: "#/components/schemas/MonitorTaskInfo"
        needs_review:
          type: array
          items:
            $ref: "#/components/schemas/MonitorTaskInfo"
        in_progress:
          type: array
          items:
            $ref: "#/components/schemas/MonitorTaskInfo"
        backlog:
          type: array
          items:
            $ref: "#/components/schemas/MonitorTaskInfo"
        closed:
          type: array
          items:
            $ref: "#/components/schemas/MonitorTaskInfo"
        timestamp:
          type: string
          format: date-time
    MonitorStatsResponse:
      type: object
      required: [stats, timestamp]
      properties:
        stats:
          $ref: "#/components/schemas/MonitorStats"
        timestamp:
          type: string
          format: date-time
    MonitorSyncResponse:
      type: object
      required: [sync, timestamp]
      properties:
        sync:
          $ref: "#/components/schemas/MonitorSyncInfo"
        timestamp:
          type: string
          format: date-time
...
    UsageResponse:
      type: object
      required:
        - total_input_tokens
        - total_output_tokens
        - total_cache_read_tokens
        - total_cache_write_tokens
        - total_cost
        - session_count
        - by_agent
        - by_backend
        - daily_costs
        - sessions
        - timestamp
      properties:
        total_input_tokens:
          type: integer
          format: int64
        total_output_tokens:
          type: integer
          format: int64
        total_cache_read_tokens:
          type: integer
          format: int64
        total_cache_write_tokens:
          type: integer
          format: int64
        total_cost:
          type: number
          format: double
        session_count:
          type: integer
        by_agent:
          type: array
          items:
            $ref: "#/components/schemas/UsageAgentSummary"
        by_backend:
          type: array
          items:
            $ref: "#/components/schemas/UsageBackendSummary"
        daily_costs:
          type: array
          items:
            $ref: "#/components/schemas/UsageDailyCost"
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/UsageSessionRecord"
        timestamp:
          type: string
          format: date-time
...
    UsageDailyCost:
      type: object
      required: [date, cost, sessions]
      properties:
        date:
          type: string
        cost:
          type: number
          format: double
        sessions:
          type: integer
...
    ObservabilityEventsResponse:
      type: object
      required: [success, data, total, page, per_page]
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/ObservabilityEvent"
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
        error:
          type: string
...